# apps/api/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Service Analytics API",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/charts/revenue-daily":  {
                                                    "get":  {
                                                                "tags":  "charts",
                                                                "summary":  "Get Revenue Daily",
                                                                "description":  "Get daily revenue chart data from fact_sales (completed sales only).",
                                                                "operationId":  "get_revenue_daily_api_charts_revenue_daily_get",
                                                                "security":  "",
                                                                "parameters":  " ",
                                                                "responses":  "@{200=; 422=}"
                                                            }
                                                },
                  "/api/charts/stock-current":  {
                                                    "get":  {
                                                                "tags":  "charts",
                                                                "summary":  "Get Stock Current",
                                                                "description":  "Get current stock snapshot from v_product_current_stock.",
                                                                "operationId":  "get_stock_current_api_charts_stock_current_get",
                                                                "security":  "",
                                                                "parameters":  " ",
                                                                "responses":  "@{200=; 422=}"
                                                            }
                                                },
...
                  "/debug/debug/db":  {
                                          "get":  {
                                                      "tags":  "debug",
                                                      "summary":  "Debug Db",
                                                      "description":  "Debug endpoint to check database connection and schema configuration.\nRequires authentication (JWT or X-User-Id header).",
                                                      "operationId":  "debug_db_debug_debug_db_get",
                                                      "responses":  "@{200=}"
                                                  }
                                      },
                  "/debug/debug/period":  {
                                              "get":  {
                                                          "tags":  "debug",
                                                          "summary":  "Debug Period",
                                                          "description":  "Debug endpoint to check period range calculation.\nReturns: user_id, period, shop_id, period_range, data_end_date, year_start, price_window_from.",
                                                          "operationId":  "debug_period_debug_debug_period_get",
                                                          "parameters":  " ",
                                                          "responses":  "@{200=; 422=}"
                                                      }
                                          },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "HTTPBearer":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              }
                                           }
                   }
}